refactor: moved getLeft & getRight styles to utils
